feat(sbt-plugin): support generating configs for all LLM clients via 'all' parameter
